+++ src/py/mat3ra/made/tools/utils/coordinate.py
@@ -1,5 +1,5 @@
 # Place all functions acting on coordinates
-from typing import List
+from typing import List, Callable, Tuple, Dict
 
 import numpy as np
 
@@ -152,3 +152,74 @@ def is_coordinate_behind_plane(
     np_plane_normal = np.array(plane_normal)
     np_plane_point = np.array(plane_point_coordinate)
     return np.dot(np_plane_normal, np_coordinate - np_plane_point) < 0
+
+
+class CoordinateConditionBuilder:
+    @staticmethod
+    def create_condition(condition_type: str, evaluation_func: Callable, **kwargs) -> Tuple[Callable, Dict]:
+        condition_json = {"type": condition_type, **kwargs}
+        return lambda coordinate: evaluation_func(coordinate, **kwargs), condition_json
+
+    @staticmethod
+    def cylinder(center_position=None, radius: float = 0.25, min_z: float = 0, max_z: float = 1):
+        if center_position is None:
+            center_position = [0.5, 0.5]
+        return CoordinateConditionBuilder.create_condition(
+            condition_type="cylinder",
+            evaluation_func=is_coordinate_in_cylinder,
+            center_position=center_position,
+            radius=radius,
+            min_z=min_z,
+            max_z=max_z,
+        )
+
+    @staticmethod
+    def sphere(center_position=None, radius: float = 0.25):
+        if center_position is None:
+            center_position = [0.5, 0.5, 0.5]
+        return CoordinateConditionBuilder.create_condition(
+            condition_type="sphere",
+            evaluation_func=is_coordinate_in_sphere,
+            center_position=center_position,
+            radius=radius,
+        )
+
+    @staticmethod
+    def triangular_prism(
+        position_on_surface_1: List[float] = [0, 0],
+        position_on_surface_2: List[float] = [1, 0],
+        position_on_surface_3: List[float] = [0, 1],
+        min_z: float = 0,
+        max_z: float = 1,
+    ):
+        return CoordinateConditionBuilder.create_condition(
+            condition_type="prism",
+            evaluation_func=is_coordinate_in_triangular_prism,
+            coordinate_1=position_on_surface_1,
+            coordinate_2=position_on_surface_2,
+            coordinate_3=position_on_surface_3,
+            min_z=min_z,
+            max_z=max_z,
+        )
+
+    @staticmethod
+    def box(min_coordinate=None, max_coordinate=None):
+        if max_coordinate is None:
+            max_coordinate = [1, 1, 1]
+        if min_coordinate is None:
+            min_coordinate = [0, 0, 0]
+        return CoordinateConditionBuilder.create_condition(
+            condition_type="box",
+            evaluation_func=is_coordinate_in_box,
+            min_coordinate=min_coordinate,
+            max_coordinate=max_coordinate,
+        )
+
+    @staticmethod
+    def plane(plane_normal: List[float], plane_point_coordinate: List[float]):
+        return CoordinateConditionBuilder.create_condition(
+            condition_type="plane",
+            evaluation_func=is_coordinate_behind_plane,
+            plane_normal=plane_normal,
+            plane_point_coordinate=plane_point_coordinate,
+        )
