+++ tests/py/unit/test_tools_build_defect.py
@@ -19,7 +19,7 @@ from mat3ra.made.tools.build.defect.configuration import (
     PointDefectPairConfiguration,
     TerraceSlabDefectConfiguration,
 )
-from mat3ra.made.tools.utils import CoordinateConditionBuilder
+from mat3ra.made.tools.utils.coordinate import CoordinateConditionBuilder
 from mat3ra.utils import assertion as assertion_utils
 
 from .fixtures import SLAB_001, SLAB_111
